% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{cluster_graph}
\alias{cluster_graph}
\title{run umap algorithm on igg data}
\usage{
cluster_graph(knn_graph, cluster_method = "louvain")
}
\arguments{
\item{knn_graph}{igraph object}

\item{cluster_method}{graph cluster method. Default = "louvain"}
}
\value{
igraph cluster object
}
\description{
run umap algorithm on igg data
}
