% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{compare_clusters_alluvial}
\alias{compare_clusters_alluvial}
\title{Alluvial plot to compare igraph cluster objects}
\usage{
compare_clusters_alluvial(
  cluster1 = daisy_clusters,
  cluster2 = pagoda_clusters
)
}
\arguments{
\item{cluster1}{igraph cluster object 1}

\item{cluster2}{igraph cluster object 2}
}
\value{
alluvial plot object and table of membership
}
\description{
Alluvial plot to compare igraph cluster objects
}
