% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph_clusters.R
\name{convert_to_ggplotformat}
\alias{convert_to_ggplotformat}
\title{convert model stats to ggplot format}
\usage{
convert_to_ggplotformat(daisy_models = daisy_eigen_models$fulltable)
}
\arguments{
\item{daisy_models}{output from ??univariate_models}
}
\value{
table of linear model results, including effect size and 95\% CI
}
\description{
This function converts the output from ??univariate_models to ggplot2 format
to prepare for using ??forest_plot_lm_model function
}
