% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{entropy}
\alias{entropy}
\title{entropy/shannon diversity index function}
\usage{
entropy(pop)
}
\arguments{
\item{pop}{vector}
}
\value{
entropy value
}
\description{
This function calculates the entropy of shannon diversity index for a given vector
}
