% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph_clusters.R
\name{forest_plot_lm_model}
\alias{forest_plot_lm_model}
\title{plot ggplot converted lm model as forest plot}
\usage{
forest_plot_lm_model(daisy_model)
}
\arguments{
\item{daisy_model}{output from ??convert_to_ggplotformat}
}
\value{
forest plot as a ggplot2 object
}
\description{
This function plots the results from ??convert_to_ggplotformat as a forest plot
}
