% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph_clusters.R
\name{glycan_class_enrichment}
\alias{glycan_class_enrichment}
\title{glycan class enrichment analysis}
\usage{
glycan_class_enrichment(co_cluster = co_cluster, glygroups = glycan_classes)
}
\arguments{
\item{co_cluster}{data frame of iggdata}

\item{glygroups}{p-value cutoff for edge removal from graph. Default = 0.05}
}
\value{
list including a full table and partial table of significantly enriched classes
}
\description{
This function calculates the significance of enrichment of glycans for
each glycan class in each glycan cluster using fisher's exact test
}
