% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{join_graphs}
\alias{join_graphs}
\title{combine two igraph objects}
\usage{
join_graphs(graph1 = daisy_knn_graph, graph2 = pagoda_knn_graph)
}
\arguments{
\item{graph1}{First igraph object}

\item{graph2}{Second igraph object}
}
\value{
igraph object
}
\description{
combine two igraph objects
}
