% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{keep_earliest_samples}
\alias{keep_earliest_samples}
\title{Keep earliest sample only for multiple samples}
\usage{
keep_earliest_samples(iggdata)
}
\arguments{
\item{iggdata}{data frame with repeat samples}
}
\value{
data frame  with only earliest samples
}
\description{
Keep earliest sample only for multiple samples
}
