% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{knn_network_from_umap}
\alias{knn_network_from_umap}
\title{Construct K-nearest neighbors graph from UMAP implementation of KNN}
\usage{
knn_network_from_umap(myumap, metric = "cosine")
}
\arguments{
\item{myumap}{output from make_umap function}

\item{metric}{hyperparameter distance metric used for umap function. Default = "cosine"}
}
\value{
kNN graph as an igraph object
}
\description{
Construct K-nearest neighbors graph from UMAP implementation of KNN
}
