% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_and_clean_DAISY_data}
\alias{load_and_clean_DAISY_data}
\title{load and clean DAISY iggdata}
\usage{
load_and_clean_DAISY_data(
  dataloc,
  phenoloc,
  removeoutliers = F,
  backgroundThesholdval = 1,
  log2 = F
)
}
\arguments{
\item{dataloc}{Path to DAISY Iggdata}

\item{phenoloc}{Path to DAISY phenotype data}

\item{removeoutliers}{Logical indicating whether or not to remove outliers in the iggdata}

\item{backgroundThesholdval}{Value for background thresholding of Iggdata}

\item{log2}{Logical indicating whether or not to log2 transform the iggdata}
}
\value{
a list containing two data frames, igg and pheno
}
\description{
This function loads the serum anti-carbohydrate
antibody data from the DAISY cohort along
with the DAISY phenotype data
It keeps only the earliest measured sample
for each individual and performs background
thresholding
}
