% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_and_clean_PAGODA_data}
\alias{load_and_clean_PAGODA_data}
\title{load and clean PAGODA iggdata}
\usage{
load_and_clean_PAGODA_data(
  fileloc,
  removeoutliers = F,
  backgroundThesholdval = 1,
  log2 = F
)
}
\arguments{
\item{fileloc}{Path to PAGODA data}

\item{removeoutliers}{Logical indicating whether or not to remove outliers in the iggdata}

\item{backgroundThesholdval}{Value for background thresholding of Iggdata}

\item{log2}{Logical indicating whether or not to log2 transform the iggdata}
}
\value{
a list containing two data frames, igg and pheno
}
\description{
This function loads the serum anti-carbohydrate
antibody data from the PAGODA cohort along
with the PAGODA phenotype data
It keeps only the earliest measured sample
for each individual and performs background
thresholding
}
