% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_glycan_classes}
\alias{load_glycan_classes}
\title{Load glycan classes information}
\usage{
load_glycan_classes(fileloc = "/Users/paultran/Downloads/glycan_classes.csv")
}
\arguments{
\item{fileloc}{Path to location of glycan class data file}
}
\value{
data frame for classification of glycans
}
\description{
This function loads the classification information for glycans
}
