% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph_clusters.R
\name{make_correlation_network}
\alias{make_correlation_network}
\title{Construct igraph based on pairwise spearman correlations}
\usage{
make_correlation_network(
  iggdata,
  pcutoff = 0.05,
  rcutoff = 0.5,
  n_neighbor = 2
)
}
\arguments{
\item{iggdata}{data frame of iggdata}

\item{pcutoff}{p-value cutoff for edge removal from graph. Default = 0.05}

\item{rcutoff}{Spearman's rho cutoff for edge removal from graph. Default = 0.5}

\item{n_neighbor}{keep edges based on number of nearest neighbors. Default = 2}
}
\value{
igraph object
}
\description{
This function is not recommended since a K-nearest neighbors approach
with louvain clustering is a more generalizable approach to clustering.
See the ??make_umap and ??knn_network_from_umap functions for more details.
This function calculates the pairwise spearman correlaton
for all anti-carbohydrate antibodies, then constructs a graph based
on the spearman's rho and prunes weak edges based on p-value and
spearmna's rho cutoffs.
}
