% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{make_eigen_matrix}
\alias{make_eigen_matrix}
\title{Make eigen matrix}
\usage{
make_eigen_matrix(mydata, myclusters)
}
\arguments{
\item{mydata}{igg data}

\item{myclusters}{vector indicating cluster membership for each variable}
}
\value{
data frame with rows as samples and columns as PC1 of each cluster
}
\description{
for each cluster of variables, calculate the first
principle component the cluster
}
