% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{make_umap}
\alias{make_umap}
\title{run umap algorithm on igg data}
\usage{
make_umap(
  datamat,
  n_neighbors = 15,
  min_dist = 0.1,
  metric = "cosine",
  random_state = 123
)
}
\arguments{
\item{datamat}{data matrix containing igg data}

\item{n_neighbors}{number of neighbors to use in umap hyperparameters. Default = 15}

\item{min_dist}{hyperparameter minimum distance for umap. Default = 0.1}

\item{metric}{hyperparameter distance metric to use for umap. Default = "cosine"}

\item{random_state}{umap variable for reproducible results. Default = 123}
}
\value{
umap object including k-nearest neighbors results and 2D projection
}
\description{
run umap algorithm on igg data
}
