% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{outliers_to_na}
\alias{outliers_to_na}
\title{outliers to na}
\usage{
outliers_to_na(iggdata)
}
\arguments{
\item{iggdata}{numeric matrix}
}
\value{
numeric matrix with outliers converted to "NA"
}
\description{
This function takes a data frame and converts all column-wise outliers to na
An outlier is a value less than the 1st quartile minus the interquartile range
or a value greater than the 3rd quartile plus the interquartile range
}
