% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph_clusters.R
\name{plot_significant_cluster_classes}
\alias{plot_significant_cluster_classes}
\title{plot result of glycan enrichment analysis as bubble chart on ggplot2}
\usage{
plot_significant_cluster_classes(cluster_class_enrichment)
}
\arguments{
\item{cluster_class_enrichment}{output from ??glycan_class_enrichment}
}
\value{
bubble chart
}
\description{
This function plots the results from ??glycan_class_enrichment as a bubble chart
}
