% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph_clusters.R
\name{univariate_models}
\alias{univariate_models}
\title{regression models}
\usage{
univariate_models(newdatafile, groups)
}
\arguments{
\item{newdatafile}{data frame with all independent and dependent variables to be modeled}

\item{groups}{a vector of all dependent variables. Example: groups<-c("Group", "Sex", "FDR", "HLA_risk","Draw_Age")}
}
\value{
list object including abbreviated summary table of significant associations, summary table of all associations, and all models
}
\description{
This function performs linear regression for each column in a given data frame
except for those listed as a dependent variable
}
