% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{univariate_radar_chart}
\alias{univariate_radar_chart}
\title{rader/spider chart per glycan class}
\usage{
univariate_radar_chart(
  daisy = daisy,
  gly_class = "Aminoglycoside antibiotics",
  label = T
)
}
\arguments{
\item{daisy}{igg data}

\item{gly_class}{glycan class to assess. Default = "Aminoglycoside antibiotics"}
}
\value{
spider plot
}
\description{
This function plots the significance of association for each anti-carbohydrate
antibody in a class against islet autoimmunity and progression to type 1 diabetes
accounting for covariates sex, first-degree relative, HLA risk, and age. This uses the
??univariate_models function
}
